Add files via upload
ABSTRACT: Construction is the process of constructing a building or
infrastructure. Construction differs from manufacturing. Manufacturing
typically involves mass production of similar items without a designated
purchaser, while construction typically takes place on location for a known
client. Construction is directly tied to the fields of civil engineering and
architecture. A construction company is responsible for building structures in
the commercial and private sectors. In simple words, we can say that a
construction company is a type of business, enterprise, or similar organization
created and operating to construct a wide variety of buildings, developments,
housing, path, pavement, roads, motorways, and other types of construction
projects. A construction company involves lot of parameters like details of
projects, employees, machineries and raw materials. A good management
system will reduce the tedious work of managing various factors in a
Construction company.
OBJECTIVE:
1. The company will be able to easily track the details of projects, employees,
machineries and raw materials.
2. It will give a proper relation regarding which employees are working in which
projects.
3. It will give details regarding how many raw materials are being allocated to
each project.
4. It will help in tracking the machineries linked to each project.